requirements.txt
aiogram==2.25.2
psycopg2-binary==2.9.10
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
environs==11.2.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.12.0
lxml==5.3.0
magic-filter==1.0.12
marshmallow==3.23.1
multidict==6.1.0
openai==2.8.1
packaging==24.2
pillow==12.0.0
propcache==0.2.0
pydantic==2.12.4
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
xlsxwriter==3.2.9
yarl==1.17.1
